requirements.txt
APScheduler==3.10.1
backports.zoneinfo==0.2.1
bcrypt==4.0.1
beautifulsoup4==4.11.2
blinker==1.5
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
DateTime==5.1
dnspython==2.3.0
filelock==3.10.0
Flask==2.2.3
Flask-Bcrypt==1.0.1
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-WTF==1.1.1
idna==3.4
importlib-metadata==6.1.0
itsdangerous==2.1.2
javascript==1!1.0.1
Jinja2==3.1.2
lxml==4.9.2
MarkupSafe==2.1.2
pkg_resources==0.0.0
PyJWT==2.0.0
pymongo==4.3.3
PySocks==1.7.1
python-dotenv==1.0.0
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
requests==2.28.2
six==1.16.0
snscrape==0.6.2.20230320
soupsieve==2.4
tzdata==2022.7
tzlocal==4.3
urllib3==1.26.15
uuid==1.30
Werkzeug==2.2.3
WTForms==3.0.1
zipp==3.15.0
zope.interface==6.0
